🚚 rename: 중복 폴더 파일 이동
