+++ src/main/java/com/mulmeong/comment/infrastructure/ShortsCommentRepositoryCustomImpl.java
@@ -1,7 +1,6 @@
-package com.mulmeong.comment.infrasturcture;
+package com.mulmeong.comment.infrastructure;
 
 import com.mulmeong.comment.common.utils.CursorPage;
-import com.mulmeong.comment.dto.out.ShortsCommentResponseDto;
 import com.mulmeong.comment.entity.QShortsComment;
 import com.mulmeong.comment.entity.ShortsComment;
 import com.querydsl.core.BooleanBuilder;
